+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## 1.12.2-1.1.0
+### Added
+- Added mob drop baked enchantment injection (configurable)
+- Added mob update baked enchantment injection (configurable)
+
+### Fixed
+- Fixed JEI log error spam
 
 ## 1.12.2-1.0.0
 - Initial Release